add: ProductItem 리팩토링
